+++ ethereum/executionclient/src/main/java/tech/pegasys/teku/ethereum/executionclient/methods/EngineExchangeTransitionConfigurationV1.java
@@ -33,7 +33,7 @@ public class EngineExchangeTransitionConfigurationV1
 
   @Override
   public String getName() {
-    return EngineApiMethods.ENGINE_EXCHANGE_TRANSITION_CONFIGURATION.getName();
+    return EngineApiMethod.ENGINE_EXCHANGE_TRANSITION_CONFIGURATION.getName();
   }
 
   @Override