+++ ethereum/executionclient/src/test/java/tech/pegasys/teku/ethereum/executionclient/methods/EngineGetPayloadV2Test.java
@@ -40,7 +40,6 @@ import tech.pegasys.teku.spec.TestSpecFactory;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadContext;
 import tech.pegasys.teku.spec.datastructures.execution.GetPayloadResponse;
-import tech.pegasys.teku.spec.datastructures.execution.versions.bellatrix.ExecutionPayloadBellatrix;
 import tech.pegasys.teku.spec.datastructures.execution.versions.capella.ExecutionPayloadCapella;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
 
@@ -110,43 +109,11 @@ class EngineGetPayloadV2Test {
   }
 
   @Test
-  public void shouldCallGetPayloadV2AndParseResponseSuccessfullyWhenInBellatrix() {
-    final Spec bellatrixSpec = TestSpecFactory.createMinimalBellatrix();
-    final DataStructureUtil dataStructureUtilBellatrix = new DataStructureUtil(bellatrixSpec);
-
+  public void shouldCallGetPayloadV2AndParseResponseSuccessfully() {
     final ExecutionPayloadContext executionPayloadContext =
-        dataStructureUtilBellatrix.randomPayloadExecutionContext(false);
+        dataStructureUtil.randomPayloadExecutionContext(false);
     final UInt256 blockValue = UInt256.MAX_VALUE;
-    final ExecutionPayload executionPayloadBellatrix =
-        dataStructureUtilBellatrix.randomExecutionPayload();
-    assertThat(executionPayloadBellatrix).isInstanceOf(ExecutionPayloadBellatrix.class);
-
-    when(executionEngineClient.getPayloadV2(eq(executionPayloadContext.getPayloadId())))
-        .thenReturn(dummySuccessfulResponse(executionPayloadBellatrix, blockValue));
-
-    final JsonRpcRequestParams params =
-        new JsonRpcRequestParams.Builder().add(executionPayloadContext).add(UInt64.ZERO).build();
-
-    jsonRpcMethod = new EngineGetPayloadV2(executionEngineClient, bellatrixSpec);
-
-    final GetPayloadResponse expectedGetPayloadResponse =
-        new GetPayloadResponse(executionPayloadBellatrix, blockValue);
-    assertThat(jsonRpcMethod.execute(params)).isCompletedWithValue(expectedGetPayloadResponse);
-
-    verify(executionEngineClient).getPayloadV2(eq(executionPayloadContext.getPayloadId()));
-    verifyNoMoreInteractions(executionEngineClient);
-  }
-
-  @Test
-  public void shouldCallGetPayloadV2AndParseResponseSuccessfullyWhenInCapella() {
-    final Spec capellaSpec = TestSpecFactory.createMinimalCapella();
-    final DataStructureUtil dataStructureUtilCapella = new DataStructureUtil(capellaSpec);
-
-    final ExecutionPayloadContext executionPayloadContext =
-        dataStructureUtilCapella.randomPayloadExecutionContext(false);
-    final UInt256 blockValue = UInt256.MAX_VALUE;
-    final ExecutionPayload executionPayloadCapella =
-        dataStructureUtilCapella.randomExecutionPayload();
+    final ExecutionPayload executionPayloadCapella = dataStructureUtil.randomExecutionPayload();
     assertThat(executionPayloadCapella).isInstanceOf(ExecutionPayloadCapella.class);
 
     when(executionEngineClient.getPayloadV2(eq(executionPayloadContext.getPayloadId())))
@@ -155,7 +122,7 @@ class EngineGetPayloadV2Test {
     final JsonRpcRequestParams params =
         new JsonRpcRequestParams.Builder().add(executionPayloadContext).add(UInt64.ZERO).build();
 
-    jsonRpcMethod = new EngineGetPayloadV2(executionEngineClient, capellaSpec);
+    jsonRpcMethod = new EngineGetPayloadV2(executionEngineClient, spec);
 
     final GetPayloadResponse expectedGetPayloadResponse =
         new GetPayloadResponse(executionPayloadCapella, blockValue);