+++ ethereum/executionlayer/src/test/java/tech/pegasys/teku/ethereum/executionlayer/DenebExecutionClientHandlerTest.java
@@ -14,8 +14,6 @@
 package tech.pegasys.teku.ethereum.executionlayer;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -26,24 +24,24 @@ import org.apache.tuweni.units.bigints.UInt256;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.ethereum.executionclient.schema.BlobsBundleV1;
-import tech.pegasys.teku.ethereum.executionclient.schema.ExecutionPayloadV1;
-import tech.pegasys.teku.ethereum.executionclient.schema.ExecutionPayloadV2;
 import tech.pegasys.teku.ethereum.executionclient.schema.ExecutionPayloadV3;
+import tech.pegasys.teku.ethereum.executionclient.schema.ForkChoiceStateV1;
+import tech.pegasys.teku.ethereum.executionclient.schema.ForkChoiceUpdatedResult;
 import tech.pegasys.teku.ethereum.executionclient.schema.GetPayloadV3Response;
+import tech.pegasys.teku.ethereum.executionclient.schema.PayloadAttributesV2;
 import tech.pegasys.teku.ethereum.executionclient.schema.PayloadStatusV1;
 import tech.pegasys.teku.ethereum.executionclient.schema.Response;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.TestSpecFactory;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadContext;
 import tech.pegasys.teku.spec.datastructures.execution.GetPayloadResponse;
 import tech.pegasys.teku.spec.datastructures.execution.NewPayloadRequest;
-import tech.pegasys.teku.spec.datastructures.execution.versions.bellatrix.ExecutionPayloadBellatrix;
-import tech.pegasys.teku.spec.datastructures.execution.versions.capella.ExecutionPayloadCapella;
 import tech.pegasys.teku.spec.datastructures.execution.versions.deneb.ExecutionPayloadDeneb;
 import tech.pegasys.teku.spec.executionlayer.ExecutionPayloadStatus;
+import tech.pegasys.teku.spec.executionlayer.ForkChoiceState;
+import tech.pegasys.teku.spec.executionlayer.PayloadBuildingAttributes;
 import tech.pegasys.teku.spec.executionlayer.PayloadStatus;
 import tech.pegasys.teku.spec.logic.versions.deneb.types.VersionedHash;
 import tech.pegasys.teku.spec.util.DataStructureUtil;
@@ -57,75 +55,7 @@ public class DenebExecutionClientHandlerTest extends ExecutionHandlerClientTest 
   }
 
   @Test
-  void engineGetPayload_bellatrixFork() throws ExecutionException, InterruptedException {
-    final UInt64 bellatrixForkEpoch = UInt64.valueOf(0);
-    final UInt64 capellaForkEpoch = UInt64.valueOf(1);
-    final UInt64 denebForkEpoch = UInt64.valueOf(2);
-    final Spec denebSpecWithForkSchedule =
-        TestSpecFactory.createMinimalWithCapellaAndDenebForkEpoch(capellaForkEpoch, denebForkEpoch);
-
-    final ExecutionClientHandler handler =
-        new ExecutionClientHandlerImpl(
-            new MilestoneBasedExecutionJsonRpcMethodsResolver(
-                denebSpecWithForkSchedule,
-                new LocalEngineApiCapabilitiesProvider(
-                    denebSpecWithForkSchedule, executionEngineClient)));
-    final DataStructureUtil data = new DataStructureUtil(denebSpecWithForkSchedule);
-    final ExecutionPayloadContext context = randomContext();
-
-    final SafeFuture<Response<GetPayloadV3Response>> dummyResponse =
-        SafeFuture.completedFuture(
-            new Response<>(
-                new GetPayloadV3Response(
-                    ExecutionPayloadV3.fromInternalExecutionPayload(data.randomExecutionPayload()),
-                    dataStructureUtil.randomUInt256(),
-                    null)));
-    when(executionEngineClient.getPayloadV3(context.getPayloadId())).thenReturn(dummyResponse);
-
-    final UInt64 bellatrixSlot =
-        denebSpecWithForkSchedule.computeStartSlotAtEpoch(bellatrixForkEpoch);
-    final SafeFuture<GetPayloadResponse> future = handler.engineGetPayload(context, bellatrixSlot);
-    verify(executionEngineClient).getPayloadV3(context.getPayloadId());
-    verify(executionEngineClient, never()).getPayloadV2(any());
-    verify(executionEngineClient, never()).getPayloadV1(any());
-
-    assertThat(future).isCompleted();
-    assertThat(future.get().getExecutionPayload()).isInstanceOf(ExecutionPayloadBellatrix.class);
-  }
-
-  @Test
-  void engineGetPayload_capellaFork() throws ExecutionException, InterruptedException {
-    final UInt64 capellaForkEpoch = UInt64.valueOf(0);
-    final UInt64 denebForkEpoch = UInt64.valueOf(1);
-    final Spec denebSpecWithForkSchedule =
-        TestSpecFactory.createMinimalWithDenebForkEpoch(denebForkEpoch);
-    final ExecutionClientHandler handler =
-        new ExecutionClientHandlerImpl(
-            new MilestoneBasedExecutionJsonRpcMethodsResolver(
-                denebSpecWithForkSchedule,
-                new LocalEngineApiCapabilitiesProvider(
-                    denebSpecWithForkSchedule, executionEngineClient)));
-    final DataStructureUtil data = new DataStructureUtil(denebSpecWithForkSchedule);
-    final ExecutionPayloadContext context = randomContext();
-
-    final SafeFuture<Response<GetPayloadV3Response>> dummyResponse =
-        SafeFuture.completedFuture(
-            new Response<>(
-                new GetPayloadV3Response(
-                    ExecutionPayloadV3.fromInternalExecutionPayload(data.randomExecutionPayload()),
-                    UInt256.MAX_VALUE,
-                    null)));
-    when(executionEngineClient.getPayloadV3(context.getPayloadId())).thenReturn(dummyResponse);
-
-    final UInt64 capellaSlot = denebSpecWithForkSchedule.computeStartSlotAtEpoch(capellaForkEpoch);
-    final SafeFuture<GetPayloadResponse> future = handler.engineGetPayload(context, capellaSlot);
-    verify(executionEngineClient).getPayloadV3(context.getPayloadId());
-    assertThat(future).isCompleted();
-    assertThat(future.get().getExecutionPayload()).isInstanceOf(ExecutionPayloadCapella.class);
-  }
-
-  @Test
-  void engineGetPayload_denebFork() throws ExecutionException, InterruptedException {
+  void engineGetPayload_shouldCallGetPayloadV3() throws ExecutionException, InterruptedException {
     final ExecutionClientHandler handler = getHandler();
     final ExecutionPayloadContext context = randomContext();
     final SafeFuture<Response<GetPayloadV3Response>> dummyResponse =
@@ -146,71 +76,11 @@ public class DenebExecutionClientHandlerTest extends ExecutionHandlerClientTest 
   }
 
   @Test
-  void engineNewPayload_bellatrixFork() {
-    final UInt64 capellaForkEpoch = UInt64.valueOf(1);
-    final UInt64 denebForkEpoch = UInt64.valueOf(2);
-    final Spec denebSpecStartingAtBellatrix =
-        TestSpecFactory.createMinimalWithCapellaAndDenebForkEpoch(capellaForkEpoch, denebForkEpoch);
-
-    final ExecutionClientHandler handler =
-        new ExecutionClientHandlerImpl(
-            new MilestoneBasedExecutionJsonRpcMethodsResolver(
-                denebSpecStartingAtBellatrix,
-                new LocalEngineApiCapabilitiesProvider(
-                    denebSpecStartingAtBellatrix, executionEngineClient)));
-    final DataStructureUtil data = new DataStructureUtil(denebSpecStartingAtBellatrix);
-
-    final ExecutionPayload payload = data.randomExecutionPayload(UInt64.ONE);
-    final NewPayloadRequest newPayloadRequest = new NewPayloadRequest(payload);
-    final ExecutionPayloadV1 payloadV1 = ExecutionPayloadV1.fromInternalExecutionPayload(payload);
-    final SafeFuture<Response<PayloadStatusV1>> dummyResponse =
-        SafeFuture.completedFuture(
-            new Response<>(
-                new PayloadStatusV1(ExecutionPayloadStatus.ACCEPTED, data.randomBytes32(), null)));
-    when(executionEngineClient.newPayloadV3(payloadV1, Optional.empty())).thenReturn(dummyResponse);
-    final SafeFuture<PayloadStatus> future = handler.engineNewPayload(newPayloadRequest);
-    verify(executionEngineClient).newPayloadV3(payloadV1, Optional.empty());
-    verify(executionEngineClient, never()).newPayloadV2(payloadV1);
-    verify(executionEngineClient, never()).newPayloadV1(payloadV1);
-    assertThat(future).isCompleted();
-  }
-
-  @Test
-  void engineNewPayload_capellaFork() {
-    final Spec denebSpecStartingAtCapella =
-        TestSpecFactory.createMinimalWithDenebForkEpoch(UInt64.ONE);
-
-    final ExecutionClientHandler handler =
-        new ExecutionClientHandlerImpl(
-            new MilestoneBasedExecutionJsonRpcMethodsResolver(
-                denebSpecStartingAtCapella,
-                new LocalEngineApiCapabilitiesProvider(
-                    denebSpecStartingAtCapella, executionEngineClient)));
-    final DataStructureUtil data = new DataStructureUtil(denebSpecStartingAtCapella);
-
-    final ExecutionPayload payload = data.randomExecutionPayload(UInt64.ONE);
-    final NewPayloadRequest newPayloadRequest = new NewPayloadRequest(payload);
-    final ExecutionPayloadV2 payloadV2 = ExecutionPayloadV2.fromInternalExecutionPayload(payload);
-    final SafeFuture<Response<PayloadStatusV1>> dummyResponse =
-        SafeFuture.completedFuture(
-            new Response<>(
-                new PayloadStatusV1(ExecutionPayloadStatus.ACCEPTED, data.randomBytes32(), null)));
-    when(executionEngineClient.newPayloadV3(payloadV2, Optional.empty())).thenReturn(dummyResponse);
-    final SafeFuture<PayloadStatus> future = handler.engineNewPayload(newPayloadRequest);
-    verify(executionEngineClient).newPayloadV3(payloadV2, Optional.empty());
-    verify(executionEngineClient, never()).newPayloadV2(payloadV2);
-    verify(executionEngineClient, never()).newPayloadV1(payloadV2);
-    assertThat(future).isCompleted();
-  }
-
-  @Test
-  void engineNewPayload_denebFork() {
+  void engineNewPayload_shouldCallNewPayloadV3() {
     final ExecutionClientHandler handler = getHandler();
     final ExecutionPayload payload = dataStructureUtil.randomExecutionPayload();
-    final Optional<List<VersionedHash>> versionedHashes =
-        Optional.of(dataStructureUtil.randomVersionedHashes(3));
-    final NewPayloadRequest newPayloadRequest =
-        new NewPayloadRequest(payload, versionedHashes.get());
+    final List<VersionedHash> versionedHashes = dataStructureUtil.randomVersionedHashes(3);
+    final NewPayloadRequest newPayloadRequest = new NewPayloadRequest(payload, versionedHashes);
     final ExecutionPayloadV3 payloadV3 = ExecutionPayloadV3.fromInternalExecutionPayload(payload);
     final SafeFuture<Response<PayloadStatusV1>> dummyResponse =
         SafeFuture.completedFuture(
@@ -220,6 +90,37 @@ public class DenebExecutionClientHandlerTest extends ExecutionHandlerClientTest 
     when(executionEngineClient.newPayloadV3(payloadV3, versionedHashes)).thenReturn(dummyResponse);
     final SafeFuture<PayloadStatus> future = handler.engineNewPayload(newPayloadRequest);
     verify(executionEngineClient).newPayloadV3(payloadV3, versionedHashes);
+    assertThat(future).isCompleted();
+  }
+
+  @Test
+  void engineForkChoiceUpdated_shouldCallEngineForkChoiceUpdatedV2() {
+    final ExecutionClientHandler handler = getHandler();
+    final ForkChoiceState forkChoiceState = dataStructureUtil.randomForkChoiceState(false);
+    final ForkChoiceStateV1 forkChoiceStateV1 =
+        ForkChoiceStateV1.fromInternalForkChoiceState(forkChoiceState);
+    final PayloadBuildingAttributes attributes =
+        new PayloadBuildingAttributes(
+            dataStructureUtil.randomUInt64(),
+            dataStructureUtil.randomBytes32(),
+            dataStructureUtil.randomEth1Address(),
+            Optional.empty(),
+            Optional.of(List.of()),
+            dataStructureUtil.randomUInt64());
+    final Optional<PayloadAttributesV2> payloadAttributes =
+        PayloadAttributesV2.fromInternalPayloadBuildingAttributesV2(Optional.of(attributes));
+    final SafeFuture<Response<ForkChoiceUpdatedResult>> dummyResponse =
+        SafeFuture.completedFuture(
+            new Response<>(
+                new ForkChoiceUpdatedResult(
+                    new PayloadStatusV1(
+                        ExecutionPayloadStatus.ACCEPTED, dataStructureUtil.randomBytes32(), ""),
+                    dataStructureUtil.randomBytes8())));
+    when(executionEngineClient.forkChoiceUpdatedV2(forkChoiceStateV1, payloadAttributes))
+        .thenReturn(dummyResponse);
+    final SafeFuture<tech.pegasys.teku.spec.executionlayer.ForkChoiceUpdatedResult> future =
+        handler.engineForkChoiceUpdated(forkChoiceState, Optional.of(attributes));
+    verify(executionEngineClient).forkChoiceUpdatedV2(forkChoiceStateV1, payloadAttributes);
     assertThat(future).isCompleted();
   }